Update to latest c-algorithms.

+++ src/algo/hash-pointer.c
@@ -18,12 +18,14 @@ CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
  */
 
+#include <limits.h>
+
 #include "hash-pointer.h"
 
 /* Hash function for a generic pointer */
 
-unsigned long irmo_pointer_hash(void *location)
+unsigned int irmo_pointer_hash(void *location)
 {
-	return (unsigned long) location;
+	return (unsigned int) (unsigned long) location;
 }
